requirements/base.txt
# Base
pytz==2021.3
python-slugify==5.0.2
Pillow==8.4.0
psycopg2==2.9.2 --no-binary psycopg2

# Django
django==3.2.10

# Environment
django-environ==0.8.1

# Django Rest Framework
djangorestframework==3.13.1

# JWT
pyjwt==2.3.0

# Passwords security
argon2-cffi==21.3.0

# Static files
whitenoise==5.3.0

# Celery
redis>=4.0.2
django-redis==5.2.0
celery==5.2.1
flower==1.0.0
